% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-utilities.R
\name{add_alpha}
\alias{add_alpha}
\title{Add transparency to colours}
\usage{
add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{a vector of colours}

\item{alpha}{transparency wanted}
}
\value{
a vector of colours with added transparency
}
\description{
Add transparency to colours
}
