% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrageneric_level.R
\name{add_infrageneric_level}
\alias{add_infrageneric_level}
\title{Add infrageneric_level column to Botanic garden database.}
\usage{
add_infrageneric_level(
  data,
  TaxonNameColumn = "TaxonName",
  POWO_TaxonNameColumn = NULL,
  progress_bar = FALSE
)
}
\arguments{
\item{data}{BG database}

\item{TaxonNameColumn}{The name of the column containing the (original) Taxon name.}

\item{POWO_TaxonNameColumn}{The name of the column containing the POWO (WCVP) Taxon name.}

\item{progress_bar}{Logical flag, if TRUE show progress bar, else no progress bar}
}
\value{
The BG database with a new column called infrageneric_level
}
\description{
Add infrageneric_level column to Botanic garden database.
}
