% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{add_splitter}
\alias{add_splitter}
\title{add_splitter()}
\usage{
add_splitter(taxon_names, taxon_authors, wcvp)
}
\arguments{
\item{taxon_names}{taxon names}

\item{taxon_authors}{taxon names with author}

\item{wcvp}{enrich information}
}
\value{
a list with match and message
}
\description{
add_splitter()
}
