% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year.R
\name{add_status_year}
\alias{add_status_year}
\title{Add status_year column to Botanic garden database.}
\usage{
add_status_year(data, ItemStatusDate_column = "ItemStatusDate")
}
\arguments{
\item{data}{BG database}

\item{ItemStatusDate_column}{The name of the column containing the Taxon name.}
}
\value{
The BG database with a new column called status_year.
}
\description{
Add status_year column to Botanic garden database.
}
