% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_to_wcvp.R
\name{additional_wcvp_matching}
\alias{additional_wcvp_matching}
\title{Pick the best record given an extract of World Checklist of Vascular Plants (WCVP) database}
\usage{
additional_wcvp_matching(enrich_database_extract, message = "")
}
\arguments{
\item{enrich_database_extract}{Extract of database (often records with identical taxonomic names).}

\item{message}{Matching message.}
}
\value{
a list of length 3 containing,
\itemize{
\item \verb{$row} The row/s index of \code{enrich_database_extract} which corresponds to the best record we match to.
\item \verb{$message} A message detailing how the matched record was decided.
}
}
\description{
Pick the best record given an extract of World Checklist of Vascular Plants (WCVP) database
}
