% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{author_check}
\alias{author_check}
\title{author_check()}
\usage{
author_check(original_author, proposed_author)
}
\arguments{
\item{original_author}{original_author}

\item{proposed_author}{proposed_author}
}
\value{
message of author comparision
}
\description{
author_check()
}
