% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{author_from_taxon_name_full}
\alias{author_from_taxon_name_full}
\title{author_from_taxon_names_full()}
\usage{
author_from_taxon_name_full(taxon_names, taxon_names_full)
}
\arguments{
\item{taxon_names}{taxon names}

\item{taxon_names_full}{taxon names with author}
}
\value{
only author
}
\description{
author_from_taxon_names_full()
}
