% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{author_words}
\alias{author_words}
\title{author_words()}
\usage{
author_words(author)
}
\arguments{
\item{author}{Author that wants splitting into words.}
}
\value{
author words
}
\description{
author_words()
}
