% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSmartR-package.R
\docType{package}
\name{BGSmartR-package}
\alias{BGSmartR}
\alias{BGSmartR-package}
\title{BGSmartR: Analyse Botanic Garden collections}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for matching collection records to external datasets and creating html reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jake-powell.github.io/BGsmartR/}
}

}
\author{
\strong{Maintainer}: Jake Powell \email{jjp68@cam.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
