% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{check_taxon_typo}
\alias{check_taxon_typo}
\title{check_taxon_typo()}
\usage{
check_taxon_typo(taxon, wcvp)
}
\arguments{
\item{taxon}{The taxon_name}

\item{wcvp}{POWO database}
}
\value{
a potential fixed name, or NA if no correction is found.
}
\description{
This function checks for typos in the taxon name and returns (the first) record found in wcvp with a single change.
}
\details{
In particular, the function compares the given taxon with a single extra letter, changing one letter or removing one letter against names known to be in POWO.
}
