% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_names_authors}
\alias{clean_names_authors}
\title{clean_names_authors()}
\usage{
clean_names_authors(taxon_names, taxon_authors = NA, taxon_name_full = NA)
}
\arguments{
\item{taxon_names}{taxon names}

\item{taxon_authors}{taxon_authors}

\item{taxon_name_full}{taxon names with author}
}
\value{
list of cleaned taxon names and author
}
\description{
clean_names_authors()
}
