% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_names_authors_report}
\alias{clean_names_authors_report}
\title{clean_names_authors_report}
\usage{
clean_names_authors_report(
  original_report,
  taxon_name_col = "TaxonName",
  taxon_name_full_col = NA,
  taxon_author_col = NA
)
}
\arguments{
\item{original_report}{original_report}

\item{taxon_name_col}{taxon_name_col}

\item{taxon_name_full_col}{taxon_name_full_col}

\item{taxon_author_col}{taxon_author_col}
}
\value{
list of cleaned taxon names and author
}
\description{
clean_names_authors_report
}
