% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{collection_example}
\alias{collection_example}
\title{Example collection data}
\format{
An object of class \code{data.frame} with 123 rows and 8 columns.
}
\usage{
collection_example
}
\description{
A data frame containing a fake collection to be used in examples
}
\keyword{datasets}
