% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{convert_to_accepted_name}
\alias{convert_to_accepted_name}
\title{convert_to_accepted_name()}
\usage{
convert_to_accepted_name(original_match, wcvp)
}
\arguments{
\item{original_match}{a vector of indices corresponding to a match between a gardens database and POWO.}

\item{wcvp}{The cleaned POWO database}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}, where the match goes to the record with accepted status.
\verb{$message} a message detailing the match.
}
\description{
This function is used to update the match to POWO by checking for accepted plant names.
}
