% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-basic_stats.R
\name{create_basic_stats_static}
\alias{create_basic_stats_static}
\title{create_basic_stats_static()}
\usage{
create_basic_stats_static(
  enriched_report,
  min_year = NULL,
  collection = NULL,
  output_file = NULL
)
}
\arguments{
\item{enriched_report}{enriched report}

\item{min_year}{The first year to consider for retrospective review.}

\item{collection}{The name of the collection}

\item{output_file}{The file path of the output}
}
\description{
create_basic_stats_static()
}
