% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-duplication.R
\name{create_duplication_report}
\alias{create_duplication_report}
\title{Create rarity and duplication module}
\usage{
create_duplication_report(
  enriched_report,
  collection = NULL,
  min_year = 1970,
  join_bigger_than = 4,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{collection}{The name of the living collection}

\item{min_year}{The minimum year for analyses over time.}

\item{join_bigger_than}{When exploring accessions per item, etc the value at which we merge larger values for simplified pie charts. Default is 4.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{output_dir}{The output directory}
}
\description{
Degree of rarity of a plant in cultivation globally and duplication of individuals per species in a collection.
}
