% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_geography.R
\name{create_geography_report}
\alias{create_geography_report}
\title{Create a report of the geographic distribution of the LC (currently existing)}
\usage{
create_geography_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  report_kind = "static",
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  export_data = FALSE,
  output_file = NULL,
  output_dir = NULL,
  table_font_size = 10,
  ggtheme = NULL,
  separate_figure_folder = TRUE,
  value_on_fig = FALSE,
  color_binary = c("darkgray", "darkgreen"),
  palette = "Greens",
  scale_colour_distiller = function(...) ggplot2::scale_colour_distiller(palette =
    "Greens", ...),
  scale_fill_distiller = function(...) ggplot2::scale_fill_distiller(palette = "Greens",
    ...),
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched (via BGSmartR) report of a LC.}

\item{collection}{The name of the LC}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} pacakge.}

\item{report_kind}{The find of report to create, either \code{static} or \code{interactive}.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{output_file}{The file path of the native report.}

\item{output_dir}{The output directory}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{separate_figure_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{value_on_fig}{Flag (TRUE/FALSE) for whether to include values of static plots.}

\item{color_binary}{colours for the representation of species.}

\item{palette}{String colour palette for use in the interactive report.}

\item{scale_colour_distiller, scale_fill_distiller}{scales for ggplot. Default is to use viridis for discrete and greens for distiller.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}
}
\value{
renders the native (word document) or interactive report (html document)
}
\description{
This function can be used to create a report of the geographic distribution of plants in a living collection (LC). This includes exploring:
\itemize{
\item Which geographic regions are represented by species in the LC?
\item Which geographic regions are represented by Endemic species in the LC?
\item Which geographic regions are represented by Threatened species in the LC?
\item How many species are their from each regions in the LC?
\item How many Endemic species are their from each regions in the LC?
\item How many Threatened species are their from each regions in the LC?
}
}
