% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_geography.R
\name{create_geography_static}
\alias{create_geography_static}
\title{Create a report of the geographic distribution of the LC (currently existing)}
\usage{
create_geography_static(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  export_data = FALSE,
  output_file = NULL,
  table_font_size = 10,
  ggtheme = NULL,
  separate_plot_folder = TRUE,
  scale_colour_distiller = function(...) ggplot2::scale_colour_distiller(palette =
    "Greens", ...),
  scale_fill_distiller = function(...) ggplot2::scale_fill_distiller(palette = "Greens",
    ...),
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched (via BGSmartR) report of a LC.}

\item{collection}{The name of the LC}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} pacakge.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{output_file}{The file path of the native report.}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{separate_plot_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{scale_colour_distiller, scale_fill_distiller, scale_fill_discrete, scale_colour_discrete}{scales for ggplot. Default is to use viridis for discrete and greens for distiller.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}
}
\value{
renders the native report (word document)
}
\description{
This function can be used to create a report of the geographic distribution of plants in a living collection (LC). This includes exploring:
\itemize{
\item Which geographic regions are represented by species in the LC?
\item Which geographic regions are represented by Endemic species in the LC?
\item Which geographic regions are represented by Threatened species in the LC?
\item How many species are their from each regions in the LC?
\item How many Endemic species are their from each regions in the LC?
\item How many Threatened species are their from each regions in the LC?
}
}
