% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-key_taxa.R
\name{create_key_taxa_interactive}
\alias{create_key_taxa_interactive}
\title{create_key_taxa_interactive()}
\usage{
create_key_taxa_interactive(
  enriched_report,
  min_year = NULL,
  collection = NULL,
  output_file = NULL
)
}
\arguments{
\item{enriched_report}{enriched report}

\item{min_year}{The first year to consider for retrospective review.}

\item{collection}{The name of the collection}

\item{output_file}{The file path of the output}
}
\description{
create_key_taxa_interactive()
}
