% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-last_update.R
\name{create_last_update_interactive}
\alias{create_last_update_interactive}
\title{create_last_update_interactive()}
\usage{
create_last_update_interactive(
  enriched_report,
  collection = NULL,
  output_file = NULL
)
}
\arguments{
\item{enriched_report}{enriched report}

\item{collection}{The name of the collection}

\item{output_file}{The file path of the output}
}
\description{
create_last_update_interactive()
}
