% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-key_taxa.R
\name{create_native_report}
\alias{create_native_report}
\alias{create_endemic_report}
\alias{create_threatened_report}
\alias{create_trees_report}
\title{Create key collections module}
\usage{
create_native_report(
  enriched_report,
  coordinates = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  collection = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  min_year = 1970,
  output_dir = NULL,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  bru_codes = NA
)

create_endemic_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  endemic_species_per_region = NULL,
  coordinates = NA,
  output_dir = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  min_year = 1970,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL
)

create_threatened_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  detailed_IUCN_redlist = NULL,
  coordinates = NA,
  output_dir = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  min_year = 1970,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL
)

create_trees_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  coordinates = NA,
  output_dir = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  min_year = 1970,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} package.}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{collection}{The name of the living collection}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{min_year}{The minimum year for analyses over time.}

\item{output_dir}{The output directory}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{bru_codes}{A vector of the BRU level 3 codes that want to be considered as the native regions in the report.}

\item{endemic_species_per_region}{A data frame detailing the number of accepted endemic species found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}

\item{detailed_IUCN_redlist}{A data frame containing the history and geography of IUCN redlist information.}
}
\value{
html report
}
\description{
Create an overview of key collections within a living collection including threatened species, endemic species, native species and trees.
}
