% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_overview.R
\name{create_overview_report}
\alias{create_overview_report}
\title{Create overview module}
\usage{
create_overview_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  PlantClassification_filepath = NULL,
  coordinates = NA,
  output_dir = NULL,
  min_year = 1970,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{collection}{The name of the living collection}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} package.}

\item{PlantClassification_filepath}{filepath to family classification. If NULL we use the default contained within BGSmartR.}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{output_dir}{The output directory}

\item{min_year}{The minimum year for analyses over time.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}
}
\description{
The most relevant plots illustrating the general quality of a collection. This include sections on:
\itemize{
\item Composition of the collection,
\item Geographic coverage of the collection,
\item Change over time of the collection,
\item Turnover of the collection,
\item Taxonomic diversity of the collection,
\item Survival in the collection,
\item Provenance in the collection and,
\item Important accessions.
}
}
