% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-retrospective.R
\name{create_retrospective_interactive}
\alias{create_retrospective_interactive}
\title{create_retrospective_interactive()}
\usage{
create_retrospective_interactive(
  enriched_report,
  collection = NULL,
  min_year = NULL,
  collection_coords = NA,
  save_widgets = FALSE,
  create_excel = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{enriched_report}{enriched report}

\item{collection}{The name of the collection}

\item{min_year}{The earliest year we want to review}

\item{collection_coords}{The coordinates of the collection (long, lat)}

\item{save_widgets}{Logical for whether we want to create the folder 'Widgets' containing the widgets from the report.}

\item{create_excel}{Logical for whether we want to create an excel file containing the tables used to create the widgets in the report.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function can be used to create the snapshot report of a living collection. The report produces interactive graphs (bars, pies and geography) of the distribution of plants in the collection. This includes the native location of all plants in the collection and the proportion of endemic and threatened species in the collection.
}
