% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-snapshot.R
\name{create_snapshot_interactive}
\alias{create_snapshot_interactive}
\title{create_snapshot_interactive()}
\usage{
create_snapshot_interactive(
  enriched_report,
  collection = NULL,
  output_file = NULL,
  save_widgets = FALSE,
  create_excel = FALSE
)
}
\arguments{
\item{enriched_report}{enriched report}

\item{collection}{The name of the collection}

\item{output_file}{The file path of the output}

\item{save_widgets}{Logical for whether we want to create the folder 'Widgets' containing the widgets from the report.}

\item{create_excel}{Logical for whether we want to create an excel file containing the tables used to create the widgets in the report.}
}
\description{
create_snapshot_interactive()
}
