% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-key_taxa.R
\name{create_threatened_report}
\alias{create_threatened_report}
\title{create_threatened_report()}
\usage{
create_threatened_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  report_kind = "static",
  coordinates = NA,
  output_file = NULL,
  output_dir = NULL,
  export_data = FALSE,
  separate_figure_folder = TRUE,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  table_font_size = 10,
  ggtheme = NULL,
  value_on_fig = FALSE,
  recent_year = as.numeric(format(Sys.Date(), "\%Y")) - 3,
  min_year = 1970,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  color_binary = c("darkgray", "darkgreen"),
  palette = "Oranges",
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched (via BGSmartR) report of a LC.}

\item{collection}{The name of the LC}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} pacakge.}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{report_kind}{The find of report to create, either \code{static} or \code{interactive}.}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{output_file}{The file path of the native report.}

\item{output_dir}{The output directory}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{separate_figure_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{value_on_fig}{Flag (TRUE/FALSE) for whether to include values of static plots.}

\item{recent_year}{The year from which to do the recent analysis from.}

\item{min_year}{The minimum year for analyses over time.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned}{scales for ggplot. Default is to use viridis.}

\item{color_binary}{colours for the representation of species in maps.}

\item{palette}{String colour palette for use in the interactive report maps.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}
}
\description{
This function can be used to create the snapshot report of a living collection. The report produces interactive graphs (bars, pies and geography) of the distribution of plants in the collection. This includes the native location of all plants in the collection and the proportion of endemic and threatened species in the collection.
}
