% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-collection_dynamics.R
\name{create_trends_report}
\alias{create_trends_report}
\alias{create_turnover_report}
\title{Create collection dynamics module}
\usage{
create_trends_report(
  enriched_report,
  collection = NULL,
  min_year = 1970,
  coordinates = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  wgsrpd3 = NULL,
  data_type = "Accessions",
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  output_dir = NULL
)

create_turnover_report(
  enriched_report,
  collection = NULL,
  min_year = 1970,
  coordinates = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  wgsrpd3 = NULL,
  data_type = "Accessions",
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  output_dir = NULL
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{collection}{The name of the living collection}

\item{min_year}{The minimum year for analyses over time.}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} package.}

\item{data_type}{Either "Accessions" or "Items". Specifies the quantity to use when looking at the trends of provenance, native species, endemic species and threatened species.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{output_dir}{The output directory}
}
\description{
\strong{Trends}: Evolution over time of the capacity and diversity of a collection. \strong{Turnover}: Dynamics of the gained and lost plants each year.
}
