% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-collection_dynamics.R
\name{create_turnover_report}
\alias{create_turnover_report}
\title{create_turnover_report()}
\usage{
create_turnover_report(
  enriched_report,
  collection = NULL,
  min_year = 1970,
  coordinates = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  wgsrpd3 = NULL,
  report_kind = "static",
  data_type = "Accessions",
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  do_geography_trend = TRUE,
  output_file = NULL,
  output_dir = NULL,
  export_data = FALSE,
  save_excel = FALSE,
  separate_figure_folder = TRUE,
  table_font_size = 10,
  ggtheme = NULL,
  value_on_fig = FALSE,
  interactive_colour = viridis::viridis,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched (via BGSmartR) report of a LC.}

\item{collection}{The name of the LC}

\item{min_year}{The minimum year for analyses over time.}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} pacakge.}

\item{report_kind}{The find of report to create, either \code{static} or \code{interactive}.}

\item{data_type}{Either "Accessions" or "Items". Specifies the quantity to use when looking at the trends of provenance, native species, endemic species and threatened species.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{do_geography_trend}{Flag (TRUE/FALSE), if true we export an excel file of the trends for each  TDWG geographical codes (Brummitt, 2001) expressed to that system’s third level (369 regions). Also, if \code{separate_figure_folder == TRUE} we create a folder within \code{separate_figure_folder} called geography_trends which contains a map for each year between min_year to today. These can be joined into a video using ffmpeg e.g. ffmpeg -r 5 -i geography_trend\%d.jpeg geography_trend.mp4.}

\item{output_file}{The file path of the native report.}

\item{output_dir}{The output directory}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{save_excel}{Flag (TRUE/FALSE) if export_data == TRUE, save excel allows for a 'xlsx' version of the data to be created or not. Note that for particualarly large collections, saving an excel version may cause the function to fail.}

\item{separate_figure_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{value_on_fig}{Flag (TRUE/FALSE) for whether to include values of static plots.}

\item{interactive_colour}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned}{scales for ggplot. Default is to use viridis.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}
}
\description{
This function can be used to create the turnovers report of a living collection. The report produces interactive or static graphs (bars, pies) detailing the gain and loss of items/accessions/taxa/families over time whilst also considering the type of taxa (native, endemic, threatened) in the LC.
}
