% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_report.R
\name{enrich_collection}
\alias{enrich_collection}
\title{Enrich a collection's database}
\usage{
enrich_collection(
  collection,
  wcvp = NA,
  redList = NA,
  BGCI = NA,
  taxon_name_column = "TaxonName",
  taxon_name_full_column = NA,
  taxon_author_column = NA,
  do_is_autonym = FALSE,
  do_status_year = FALSE,
  do_taxon_types = FALSE,
  typo_method = "fast"
)
}
\arguments{
\item{collection}{A data frame containing a collection.}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{redList}{RedList database}

\item{BGCI}{Requires a cleaned BGCI plant search database to obtain how many collections globally a taxon is contained (Not freely available).}

\item{taxon_name_column}{The name of the column in the \code{collection} corresponding to taxonomic names.}

\item{taxon_name_full_column}{The name of the column in the \code{collection} corresponding to joined taxonomic names and authors.}

\item{taxon_author_column}{The name of the column in the \code{collection} corresponding to the authors of the taxonomic names.}

\item{do_is_autonym}{Flag (TRUE/FALSE) for whether to add the column is_autonym, see \code{\link[=add_is_autonym]{add_is_autonym()}} for details.}

\item{do_status_year}{Flag (TRUE/FALSE) for whether to add the column status_year, see \code{\link[=add_status_year]{add_status_year()}} for details.}

\item{do_taxon_types}{Flag (TRUE/FALSE) for whether to add the column taxon_type, see \code{\link[=add_taxon_type]{add_taxon_type()}} for details.}

\item{typo_method}{Defines the method used to find typos in the taxonomic name. Allowed values are \code{"full"},  \code{"fast"}, or \code{"no"}. These correspond to full search (see \code{\link[=match_typos]{match_typos()}}), search only using \link{typo_list} and no typo searching.}
}
\value{
The \code{collection} data frame enriched with information dependent on function inputs (new columns).
}
\description{
This function enriches plat records in a collection using information from POWO (WCVP), IUCN redlist and BGCI.
}
