% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{enrich_original}
\alias{enrich_original}
\title{enrich_original()}
\usage{
enrich_original(
  original_report,
  wcvp,
  wcvp_wanted_columns = c("plant_name_id", "taxon_name", "taxon_authors", "taxon_rank",
    "taxon_status", "powo_id", "family", "genus", "species", "lifeform_description",
    "climate_description")
)
}
\arguments{
\item{original_report}{A gardens original report}

\item{wcvp}{POWO database}

\item{wcvp_wanted_columns}{The data fields of POWO we want to enrich to the original report}
}
\value{
a list of length two:
\verb{$enriched_report} the enriched report, and
\code{match_details} the details of how taxon names have being used to match to POWO.
}
\description{
enrich_original()
}
