% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_report.R
\name{enrich_report}
\alias{enrich_report}
\title{enrich_original()}
\usage{
enrich_report(
  original_report,
  wcvp,
  do_is_autonym = FALSE,
  do_status_year = FALSE,
  do_infrageneric_level = FALSE
)
}
\arguments{
\item{original_report}{A gardens original report}

\item{wcvp}{POWO database}

\item{do_is_autonym}{Flag for whether to add the column is_autonym}

\item{do_status_year}{Flag for whether to add the column status_year}

\item{do_infrageneric_level}{Flag for whether to add the column infrageneric_level}
}
\value{
a list of length two:
\verb{$enriched_report} the enriched report, and
\code{match_details} the details of how taxon names have being used to match to POWO.
}
\description{
This function enriches an original report by adding columns:
}
\details{
\itemize{
\item \code{is_autonym} (Logical (TRUE/FALSE)) detailing whether the taxon name is an autonym.
\item \code{status_year} (numeric) the year extracted from \code{ItemStatusDate.}
\item \code{infrageneric_level} (character) the infrageneric level of the taxon name. Uses taxon names retrieved from POWO if available.
\item Information from POWO.
}
}
