% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enriched_collection_example}
\alias{enriched_collection_example}
\title{Example enriched collection data}
\format{
An object of class \code{data.frame} with 123 rows and 61 columns.
}
\usage{
enriched_collection_example
}
\description{
A data frame containing a fake enriched collection to be used in examples, corresponds to the data in "collection_example"
}
\keyword{datasets}
