% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-utilities.R
\name{exist_at_date}
\alias{exist_at_date}
\title{exist_at_date()
Function that gives the records that exist at a certain date.}
\usage{
exist_at_date(date, acc, ItemStatusDate, ItemStatusType)
}
\arguments{
\item{date}{dates for which we want the existing records.}

\item{acc}{accessions}

\item{ItemStatusDate}{ItemStatusDate}

\item{ItemStatusType}{ItemStatusType}
}
\value{
data frame where each column corresponds to a date, where the vaules is a logical vector of whether the record exists at that date.
}
\description{
exist_at_date()
Function that gives the records that exist at a certain date.
}
