% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{extract_year}
\alias{extract_year}
\title{extract_year()}
\usage{
extract_year(date_vector)
}
\arguments{
\item{date_vector}{A vector of dates}
}
\value{
a vector of years
}
\description{
A function that extracts the year from dates where the date can be in the following formats:
\itemize{
\item DD/MM/YYYY
\item YYYY-MM-DD
\item MM/YYYY
\item YYYY-MM
\item YYYY
\item DD Month YYYY
}
}
\details{
The year must be in the format YYYY. I.e. cannot have formats such as DD/MM/YY.
}
