% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{get_match_from_multiple}
\alias{get_match_from_multiple}
\title{get_match_from_multiple()}
\usage{
get_match_from_multiple(taxon_name_and_full, wcvp_mult)
}
\arguments{
\item{taxon_name_and_full}{the pair of taxon name and taxon name full.}

\item{wcvp_mult}{POWO database restricted to records that do not have a unique taxon name.}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}
\verb{$message} a message detailing the match.
}
\description{
Given a taxon name and its full version we find the records in POWO that share the same taxon name. We then choose the record to match to by:
}
\details{
\itemize{
\item Checking if the full taxon name contains the author described within POWO database. If we have a single match this is chosen.
\item If not, check for POWO records that are accepted. If a single record is accepted then this match is chosen.
\item If not, check for records that have lifeform descriptions. If we have a single record this is chosen.
\item Else, we cannot find a match and no record is chosen.
}
}
