% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{group_trend_plots_pair}
\alias{group_trend_plots_pair}
\title{Plot a pair of trend lines.}
\usage{
group_trend_plots_pair(
  years,
  data,
  name,
  report_kind,
  separate_figure_folder,
  data_type,
  split = FALSE,
  group_order = NULL,
  color_binary = interactive_colour(4, type = "trend")
)
}
\arguments{
\item{years}{years}

\item{data}{data}

\item{name}{name}

\item{report_kind}{report_kind}

\item{separate_figure_folder}{separate_figure_folder}

\item{data_type}{data_type}

\item{split}{split}

\item{group_order}{group_order}
}
\value{
plot
}
\description{
Plot a pair of trend lines.
}
