% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{import_wcvp_distribution}
\alias{import_wcvp_distribution}
\title{Import wcvp_distribution.csv}
\usage{
import_wcvp_distribution(filepath)
}
\arguments{
\item{filepath}{path to wcvp_distribution.csv file.}
}
\value{
Desired information from wcvp_distribution.csv
}
\description{
Given the path to \code{wcvp_distribution.csv} file obtained from POTW website load the relevent information into R.
}
