% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{import_wcvp_names}
\alias{import_wcvp_names}
\title{Import wcvp_names.csv}
\usage{
import_wcvp_names(filepath)
}
\arguments{
\item{filepath}{path to wcvp_names.csv file.}
}
\value{
Desired information from wcvp_names.csv
}
\description{
Given the path to \code{wcvp_names.csv} file obtained from POTW website load the relevent information into R.
}
