% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrageneric_level.R
\name{infrageneric_level}
\alias{infrageneric_level}
\title{infrageneric_level()}
\usage{
infrageneric_level(taxonName)
}
\arguments{
\item{taxonName}{taxon name}
}
\value{
The infrageneric levels of the plant
}
\description{
This functions extracts the infrageneric level from a taxon name, in particular:
}
\details{
\itemize{
\item \verb{0.indet}: The plant is indeterminate.
\item \verb{1.species}: The plant is a species.
\item \verb{2.subsp}: The plant is a subspecies.
\item \verb{3.var}: The plant is a variety.
\item \verb{4.f}: The plant is
\item \verb{5.cultivar}: The plant is a cultivar.
\item \verb{6.hybrid}: The plant is a hybrid.
}

\code{0-4} are biological characteristics and each plant can only have one of these characteristics.
\code{5-6} are horticultural characteristics and plants can have none, one or both of these.
}
\examples{
 \donttest{
# 0.indet
infrageneric_level("\xd7 Aridaria sp.")

# 1.species
infrageneric_level("Saxegothaea conspicua")

# 2.subsp
infrageneric_level("Rhododendron charitopes subsp. tsangpoense")

# 3.var
infrageneric_level("Rhododendron saluenense var. prostratum")

# 4.f
infrageneric_level("Aquilegia flabellata f. alba")
}
}
