% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{interactive_colour}
\alias{interactive_colour}
\title{Chart colours for reports}
\usage{
interactive_colour(n)
}
\arguments{
\item{n}{number of colours (allowable values = 1,2,3,4)}
}
\value{
colours
}
\description{
Chart colours for reports
}
