% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{is_autonym}
\alias{is_autonym}
\title{Find if a Taxon name is an autonym}
\usage{
is_autonym(name)
}
\arguments{
\item{name}{Taxon name of a plant}
}
\value{
TRUE if the Taxon name is an autonym, otherwise FALSE.
}
\description{
This function calculates whether a Taxon Name is an autonym.
}
\examples{
# An autonym.
is_autonym("Codiaeum variegatum var. variegatum")
# Not an autonym.
is_autonym("Crinum pedunculatum f. purple")

}
