% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_all_issue}
\alias{match_all_issue}
\title{match_all_issue()}
\usage{
match_all_issue(
  taxon_names,
  taxon_authors,
  do_taxon_author,
  wcvp,
  single_indices = NA,
  mult_indices = NA
)
}
\arguments{
\item{taxon_names}{taxon_names}

\item{taxon_authors}{taxon_authors}

\item{do_taxon_author}{do_taxon_author}

\item{wcvp}{wcvp}

\item{single_indices}{single_indices}

\item{mult_indices}{mult_indices}
}
\value{
return
}
\description{
match_all_issue()
}
