% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_to_wcvp.R
\name{match_collection_to_wcvp}
\alias{match_collection_to_wcvp}
\title{Match collection to World Checklist of Vascular Plants via taxonomic name}
\usage{
match_collection_to_wcvp(
  collection,
  wcvp,
  taxon_name_column = "TaxonName",
  taxon_name_full_column = NA,
  taxon_author_column = NA,
  typo_method = "fast",
  try_add_split = TRUE,
  try_fix_hybrid = TRUE,
  try_rm_autonym = TRUE,
  do_convert_accepted = TRUE
)
}
\arguments{
\item{collection}{A data frame containing a collection.}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{taxon_name_column}{The name of the column in the \code{collection} corresponding to taxonomic names.}

\item{taxon_name_full_column}{The name of the column in the \code{collection} corresponding to joined taxonomic names and authors.}

\item{taxon_author_column}{The name of the column in the \code{collection} corresponding to the authors of the taxonomic names.}

\item{typo_method}{Flag for whether we search for typos}

\item{try_add_split}{Flag for whether we search for missing f./var./subsp.}

\item{try_fix_hybrid}{Flag for whether we search for hybrid issues.}

\item{try_rm_autonym}{Flag for whether we try removing autonyms.}

\item{do_convert_accepted}{Flag for whether we convert to accepted names in POWO.}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}, where the match goes to the record with accepted status.
\verb{$message} a message detailing the match.
}
\description{
Match collection to World Checklist of Vascular Plants via taxonomic name
}
