% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_error}
\alias{match_error}
\title{match_error()}
\usage{
match_error(
  new_taxon_names,
  match_details,
  original_author,
  wcvp,
  current_message = ""
)
}
\arguments{
\item{new_taxon_names}{new_taxon_names}

\item{match_details}{match_details}

\item{original_author}{original_author}

\item{wcvp}{wcvp}

\item{current_message}{current_message}
}
\value{
list of message and match
}
\description{
match_error()
}
