% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_hybrid_issue}
\alias{match_hybrid_issue}
\title{match_hybrid_issue}
\usage{
match_hybrid_issue(taxon_names, taxon_names_full, wcvp)
}
\arguments{
\item{taxon_names}{taxon names}

\item{taxon_names_full}{taxon names with author}

\item{wcvp}{enrich information}
}
\value{
a list with match and message
}
\description{
match_hybrid_issue
}
