% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{match_mult_wcvp}
\alias{match_mult_wcvp}
\title{match_mult_wcvp()}
\usage{
match_mult_wcvp(taxon_names, taxon_names_full, wcvp, wcvp_search_index)
}
\arguments{
\item{taxon_names}{a vector of taxon names}

\item{taxon_names_full}{a vector of full taxon names (corresponding to \code{taxon_names})}

\item{wcvp}{cleaned POWO database}

\item{wcvp_search_index}{Indices of \code{wcvp} that we want to search for a match}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}
\verb{$message} a message detailing the match.
}
\description{
Given the taxon names and full taxon names you want to match (\code{taxon_names}, \code{taxon_names_full}), the cleaned version of plants of the world online ( \code{wcvp} imported through the function \code{import_wcvp_names()}) and the indices of \code{wcvp}  records that you want to math to (\code{wcvp_search_index}). As this function is designed to match to records that share taxon names the user should restrict the search indices to those with non-unique taxon names or a subset of those names.
}
\details{
The matching of individual taxon names is completed by \code{get_match_from_multiple()}.

The matched index corresponds to the full \code{wcvp} data set. If a match is found but we cannot ascertain which record to match to in POWO we set the match index to \code{-2}. If the taxon name is not found in POWO then the function returns \code{NA}.

Moreover, the function creates a message for each taxon name detailing when a match is found and the powo identifier and taxon name of the match in \code{wcvp}.
}
