% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_original_to_wcvp}
\alias{match_original_to_wcvp}
\title{Match report to POWO via taxon name}
\usage{
match_original_to_wcvp(
  original_report,
  wcvp,
  taxon_name_col = "TaxonName",
  taxon_name_full_col = NA,
  taxon_author_col = NA,
  typo_method = "fast",
  try_add_split = TRUE,
  try_fix_hybrid = TRUE,
  try_rm_autonym = TRUE,
  do_convert_accepted = TRUE
)
}
\arguments{
\item{original_report}{A gardens original report}

\item{wcvp}{POWO database}

\item{taxon_name_col}{Column name in the original report for taxon name}

\item{taxon_name_full_col}{Column name in the original report for taxon name and author/s (combined)}

\item{taxon_author_col}{Column name in the original report for taxon author/s}

\item{typo_method}{Flag for whether we search for typos}

\item{try_add_split}{Flag for whether we search for missing f./var./subsp.}

\item{try_fix_hybrid}{Flag for whether we search for hybrid issues.}

\item{try_rm_autonym}{Flag for whether we try removing autonyms.}

\item{do_convert_accepted}{Flag for whether we convert to accepted names in POWO.}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}, where the match goes to the record with accepted status.
\verb{$message} a message detailing the match.
}
\description{
Match report to POWO via taxon name
}
