% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_single_wcvp}
\alias{match_single_wcvp}
\title{match_single_wcvp()}
\usage{
match_single_wcvp(taxon_names, wcvp, wcvp_search_index)
}
\arguments{
\item{taxon_names}{a vector of taxon names}

\item{wcvp}{cleaned POWO database}

\item{wcvp_search_index}{Indices of \code{wcvp} that we want to search for a match}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}
\verb{$message} a message detailing the match.
}
\description{
Given the taxon names you want to match (\code{taxon_names}), the cleaned version of plants of the world online ( \code{wcvp} imported through the function \code{import_wcvp_names()}) and the indices of \code{wcvp}  records that you want to math to (\code{wcvp_search_index}). As this function is designed to match to unique taxon names the user should restrict the search indices to those with unique taxon names or a subset of those names.
}
\details{
The matched index corresponds to the full \code{wcvp} data set. If not match is found then the function returns \code{NA}.

Moreover, the function creates a message for each taxon name detailing when a match is found and the powo identifier and taxon name of the match in \code{wcvp}.
}
