% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_taxon_status}
\alias{match_taxon_status}
\title{match_taxon_status}
\usage{
match_taxon_status(author_match, corres_POWO, current_message = "")
}
\arguments{
\item{author_match}{vector of true of false whether we have a match}

\item{corres_POWO}{the corresponding entries in POWO (rows must equal length author match)}

\item{current_message}{the current message}
}
\value{
list of length 3
}
\description{
match_taxon_status
}
