% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{match_taxon_to_wcvp}
\alias{match_taxon_to_wcvp}
\title{match taxon and taxon name full to POWO}
\usage{
match_taxon_to_wcvp(taxon_taxon_full, wcvp)
}
\arguments{
\item{taxon_taxon_full}{data frame of two columns where the first column is the taxon_name and second column is the full taxon name (with authors).}

\item{wcvp}{POWO database}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}, where the match goes to the record with accepted status.
\verb{$message} a message detailing the match.
}
\description{
For each unique taxon name / taxon name full find the corresponding record in POWO.
}
