% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_typos}
\alias{match_typos}
\title{match_typos()}
\usage{
match_typos(
  taxon_names,
  taxon_names_full,
  wcvp,
  single_indices = NA,
  mult_indices = NA,
  typo_method = "fast"
)
}
\arguments{
\item{taxon_names}{taxon names}

\item{taxon_names_full}{taxon names with author}

\item{wcvp}{enrich information}

\item{single_indices}{wcvp indices with a unique taxon name}

\item{mult_indices}{wcvp indices with non-unique taxon name}

\item{typo_method}{method for finding typos}
}
\value{
a list with match and message
}
\description{
match_typos()
}
