% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_criterions.R
\name{no_additional_matching}
\alias{no_additional_matching}
\title{Blank Matching criterion}
\usage{
no_additional_matching(enrich_database_extract, message = "")
}
\arguments{
\item{enrich_database_extract}{extract of database (often records with identical taxonomic names)}

\item{message}{Matching message.}
}
\value{
Always returns a list of 3 with:
\itemize{
\item \verb{$row} =\code{-2} set to no match.
\item \verb{$message} A combination of the input message and \verb{unclear, do not match'}.
}
}
\description{
Blank Matching criterion
}
