% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{plots_turnover}
\alias{plots_turnover}
\title{plot turnover graphs}
\usage{
plots_turnover(
  turnover,
  trend,
  report_kind,
  text = "",
  separate_figure_folder,
  data_type = "items"
)
}
\arguments{
\item{turnover}{turnover}

\item{trend}{trend}

\item{report_kind}{type of report (i.e either ggplot graphs or plotly graphs)}

\item{text}{text}

\item{separate_figure_folder}{separate_figure_folder}

\item{data_type}{data_type}
}
\value{
list containing the plots
}
\description{
plot turnover graphs
}
