% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{proportional_plots_turnover}
\alias{proportional_plots_turnover}
\title{plot proportional turnover graphs}
\usage{
proportional_plots_turnover(
  turnover,
  overall_turnover,
  collection_proportion,
  separate_figure_folder,
  report_kind,
  text = "",
  data_type = "items",
  quantity = ""
)
}
\arguments{
\item{turnover}{turnover}

\item{overall_turnover}{overall_turnover}

\item{collection_proportion}{collection_proportion}

\item{separate_figure_folder}{separate_figure_folder}

\item{report_kind}{report_kind}

\item{text}{text}

\item{data_type}{data_type}

\item{quantity}{quantity}
}
\value{
list containing the plots
}
\description{
plot proportional turnover graphs
}
