% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sanitise_name}
\alias{sanitise_name}
\title{sanitise_name()}
\usage{
sanitise_name(taxon_name)
}
\arguments{
\item{taxon_name}{taxon name}
}
\value{
sanitised taxon name
}
\description{
Converts taxon names such that only the first letter of the Genus is capitalised and all remaining characters are lower case.
}
\details{
Also change to correct hybrid symbol where needed.
}
\examples{
sanitise_name('TRIGONELLA smyrnaea')
}
