% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{shorten_message}
\alias{shorten_message}
\alias{known_not_in_wcvp}
\alias{match_single_wcvp}
\alias{match_taxon_status}
\alias{get_match_from_multiple}
\alias{match_mult_wcvp}
\alias{match_rm_autonym}
\alias{match_error}
\alias{add_splitter}
\alias{match_hybrid_issue}
\alias{convert_to_accepted_name}
\alias{check_taxon_typo}
\alias{match_typos}
\alias{match_all_issue}
\title{Matching functions}
\usage{
shorten_message(messages)

known_not_in_wcvp(taxon_names)

match_single_wcvp(taxon_names, wcvp, wcvp_search_index)

match_taxon_status(author_match, corres_POWO, current_message = "")

get_match_from_multiple(taxon_name_and_author, wcvp_mult)

match_mult_wcvp(
  taxon_names,
  taxon_authors,
  wcvp,
  wcvp_search_index,
  show_progress = TRUE
)

match_rm_autonym(
  taxon_names,
  taxon_authors,
  wcvp,
  single_indices = NA,
  mult_indices = NA
)

match_error(
  new_taxon_names,
  match_details,
  original_author,
  wcvp,
  current_message = ""
)

add_splitter(taxon_names, taxon_authors, wcvp)

match_hybrid_issue(taxon_names, taxon_authors, wcvp)

convert_to_accepted_name(original_match, wcvp)

check_taxon_typo(taxon, wcvp = NA, typo_df = BGSmartR::typo_list, fast = T)

match_typos(
  taxon_names,
  taxon_authors,
  wcvp,
  single_indices = NA,
  mult_indices = NA,
  typo_method = "fast"
)

match_all_issue(
  taxon_names,
  taxon_authors,
  do_taxon_author,
  wcvp,
  single_indices = NA,
  mult_indices = NA,
  try_add_split = TRUE,
  try_fix_hybrid = TRUE,
  try_rm_autonym = TRUE
)
}
\arguments{
\item{messages}{messages about matching}

\item{taxon_names}{Vector of taxonomic names.}

\item{wcvp}{cleaned POWO database}

\item{wcvp_search_index}{Indices of \code{wcvp} that we want to search for a match}

\item{author_match}{vector of true of false whether we have a match}

\item{corres_POWO}{the corresponding entries in POWO (rows must equal length author match)}

\item{current_message}{the current message}

\item{taxon_name_and_author}{the pair of taxon name and taxon name full.}

\item{wcvp_mult}{POWO database restricted to records that do not have a unique taxon name.}

\item{taxon_authors}{a vector of full taxon names (corresponding to \code{taxon_names})}

\item{show_progress}{Flag for whether we show progress bar.}

\item{single_indices}{wcvp indices with a unique taxon name}

\item{mult_indices}{wcvp indices with non-unique taxon name}

\item{new_taxon_names}{new_taxon_names}

\item{match_details}{match_details}

\item{original_author}{original_author}

\item{original_match}{a vector of indices corresponding to a match between a gardens database and POWO.}

\item{taxon}{The taxon_name}

\item{typo_df}{A data frame with two columns of typos and their fixes.}

\item{fast}{A flag for whether we want to search only the typo_df (TRUE) or also search directly (FALSE)}

\item{typo_method}{method for finding typos}

\item{do_taxon_author}{do_taxon_author}

\item{try_add_split}{Flag for whether we search for missing f./var./subsp.}

\item{try_fix_hybrid}{Flag for whether we search for hybrid issues.}

\item{try_rm_autonym}{Flag for whether we try removing autonyms.}
}
\description{
Functions used to match taxonomic names from a collection to exterior databases (POWO's WCVP, IUCN Redlist)
}
