% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{single_trend_plot}
\alias{single_trend_plot}
\title{PLot a single trend over time}
\usage{
single_trend_plot(years, values, name, report_kind, separate_figure_folder)
}
\arguments{
\item{years}{years}

\item{values}{values}

\item{name}{name}

\item{report_kind}{report_kind}

\item{separate_figure_folder}{separate_figure_folder}
}
\value{
graph
}
\description{
PLot a single trend over time
}
