% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subfamily_genus_simp}
\alias{subfamily_genus_simp}
\title{Sub-family information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90 rows and 5 columns.
}
\usage{
subfamily_genus_simp
}
\description{
A data frame containing sub-family information used for creating sunburst/treemap charts of taxonomic diversity. Source WFO.
}
\keyword{datasets}
