% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtribe_genus_simp}
\alias{subtribe_genus_simp}
\title{Sub-tribe information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 168 rows and 5 columns.
}
\usage{
subtribe_genus_simp
}
\description{
A data frame containing sub-tribe information used for creating sunburst/treemap charts of taxonomic diversity. Source WFO.
}
\keyword{datasets}
