% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{taxonomic_standardisation}
\alias{taxonomic_standardisation}
\title{Title}
\usage{
taxonomic_standardisation(dataset)
}
\arguments{
\item{dataset}{data frame containing botanical garden reports}
}
\value{
data frame with taxonomic standardisation included.
}
\description{
Title
}
