% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tribe_genus_simp}
\alias{tribe_genus_simp}
\title{Tribe information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 5 columns.
}
\usage{
tribe_genus_simp
}
\description{
A data frame containing tribe information used for creating sunburst/treemap charts of taxonomic diversity. Source WFO.
}
\keyword{datasets}
