% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{turnover_items}
\alias{turnover_items}
\title{extract turnover information from enriched report (for items)}
\usage{
turnover_items(report)
}
\arguments{
\item{report}{the cleansed enriched report.}
}
\value{
data frame with columns $gain_items, $loss_items and  $net_items
}
\description{
extract turnover information from enriched report (for items)
}
