% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgsrpd3}
\alias{wgsrpd3}
\title{wgsrpd3}
\format{
\subsection{\code{wgsrpd3}}{

A data frame with 369 rows and 5 columns:
\describe{
\item{Original}{Original Taxon name with typo}
\item{Fixed}{Fixed Taxon name}
}
}
}
\usage{
wgsrpd3
}
\description{
A data frame containing BRU level 3 areas and their geometries.
}
\keyword{datasets}
