% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgsrpd3_level2_simp}
\alias{wgsrpd3_level2_simp}
\title{wgsrpd3_level2_simp}
\format{
\subsection{\code{wgsrpd3_level2_simp}}{

A data frame with 369 rows and 5 columns:
\describe{
\item{Original}{Original Taxon name with typo}
\item{Fixed}{Fixed Taxon name}
}
}
}
\usage{
wgsrpd3_level2_simp
}
\description{
A data frame containing BRU level 2 areas and their simplified geometries using rmapshaper::ms_simplify.
}
\keyword{datasets}
