% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-utilities.R
\name{add_alpha}
\alias{add_alpha}
\title{add_alpha}
\usage{
add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{a vector of colours}

\item{alpha}{transparancy wanted}
}
\value{
a vector of colours with added transparancy
}
\description{
add_alpha
}
