% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{add_is_autonym}
\alias{add_is_autonym}
\title{Add is_autonym column to Botanic garden database.}
\usage{
add_is_autonym(data, TaxonName_column = "TaxonName", progress_bar = FALSE)
}
\arguments{
\item{data}{BG database}

\item{TaxonName_column}{The name of the column containing the Taxon name.}

\item{progress_bar}{Logical flag, if TRUE show progress bar, else no progress bar}
}
\value{
The BG database with a new column called is_autonym which flags where the Taxon name is an autonym
}
\description{
Add is_autonym column to Botanic garden database.
}
