% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSmartR-package.R
\docType{package}
\name{BGSmartR-package}
\alias{BGSmartR}
\alias{BGSmartR-package}
\title{BGSmartR: Analyse Botanic Garden collections}
\description{
Provides tools for matching collection records to external datasets and creating html reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jakepowellcubg.github.io/BGR/}
}

}
\author{
\strong{Maintainer}: Jake Powell \email{jjp68@cam.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
