% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{check_taxon_typo}
\alias{check_taxon_typo}
\title{check_taxon_typo()}
\usage{
check_taxon_typo(taxon, wcvp)
}
\arguments{
\item{taxon}{The taxon_name}

\item{wcvp}{POWO database}
}
\value{
a potential fixed name.
}
\description{
This function checks for typos in the taxon name in reports and suggests corrections if they exist.
}
\details{
In particualar, the function compares the given taxon with a single extra letter or changing one letter against names known to be in POWO. If a fixed word is found in POWO we return it.
}
