% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{check_taxon_typo}
\alias{check_taxon_typo}
\title{check_taxon_typo()}
\usage{
check_taxon_typo(taxon, wcvp = NA, typo_df = BGSmartR::typo_list, fast = T)
}
\arguments{
\item{taxon}{The taxon_name}

\item{wcvp}{POWO database}

\item{typo_df}{A data frame with two columns of typos and their fixes.}

\item{fast}{A flag for whether we want to search only the typo_df (TRUE) or also search directly (FALSE)}
}
\value{
a potential fixed name, or NA if no correction is found.
}
\description{
This function checks for typos in the taxon name and returns (the first) record found in wcvp with a single change.
}
\details{
In particular, the function compares the given taxon with a single extra letter, changing one letter or removing one letter against names known to be in POWO.
}
