% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-data_health.R
\name{create_data_health_report}
\alias{create_data_health_report}
\title{Create data health module}
\usage{
create_data_health_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  PlantClassification_filepath = NULL,
  report_kind = "static",
  coordinates = NA,
  output_file = NULL,
  output_dir = NULL,
  export_data = FALSE,
  separate_figure_folder = TRUE,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  table_font_size = 10,
  ggtheme = NULL,
  value_on_fig = FALSE,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{collection}{The name of the living collection}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} package.}

\item{PlantClassification_filepath}{filepath to family classification. If NULL we use the default contained within BGSmartR.}

\item{report_kind}{The find of report to create, either \code{static} or \code{interactive}.}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{output_file}{The file path of the native report.}

\item{output_dir}{The output directory}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{separate_figure_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{value_on_fig}{Flag (TRUE/FALSE) for whether to include values of static plots.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned}{scales for ggplot. Default is to use viridis.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}
}
\description{
Inspection of the quality of the records and demonstrations of their enrichment
}
