% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-native.R
\name{create_native_report}
\alias{create_native_report}
\title{Create a report of native taxa in a LC}
\usage{
create_native_report(
  enriched_report,
  coordinates = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  collection = NULL,
  report_kind = "static",
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  min_year = 1970,
  export_data = FALSE,
  save_excel = FALSE,
  output_file = NULL,
  output_dir = NULL,
  table_font_size = 10,
  ggtheme = NULL,
  separate_figure_folder = TRUE,
  value_on_fig = FALSE,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched (via BGSmartR) report of a LC.}

\item{coordinates}{The coordinates of the collection (long, lat).}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} pacakge.}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{collection}{The name of the LC}

\item{report_kind}{The find of report to create, either \code{static} or \code{interactive}.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{min_year}{The minimum year for analyses over time.}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{save_excel}{Flag (TRUE/FALSE) if export_data == TRUE, save excel allows for a 'xlsx' version of the data to be created or not. Note that for particualarly large collections, saving an excel version may cause the function to fail.}

\item{output_file}{The file path of the native report.}

\item{output_dir}{The output directory}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{separate_figure_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{value_on_fig}{Flag (TRUE/FALSE) for whether to include values of static plots.}

\item{scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned}{scales for ggplot. Default is to use viridis.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}
}
\value{
renders the native report (word document)
}
\description{
This function can be used to create a report of the native plants in a living collection (LC). This includes exploring:
\itemize{
\item How many native species are in the LC?
\item What proportion of the LC at Kew is native and how has this varied over time?
\item Among the native plants in a LC, how many of them are endemic or redlisted?
\item Provenance of native plants in the LC
\item Threatened, native plants
\item Turnover of native plants
\item Survival of native plants in the LC
}
}
