% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-diversity.R
\name{create_taxonomic_diversity_report}
\alias{create_taxonomic_diversity_report}
\alias{create_geography_report}
\title{Create diversity module}
\usage{
create_taxonomic_diversity_report(
  enriched_report,
  collection = NULL,
  wcvp = NULL,
  PlantClassification = BGSmartR::Diversity_classification$Classification,
  min_year = 1970,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  report_kind = "static",
  output_file = NULL,
  output_dir = NULL,
  export_data = FALSE,
  save_excel = FALSE,
  separate_figure_folder = TRUE,
  table_font_size = 10,
  ggtheme = NULL,
  value_on_fig = FALSE,
  interactive_colour = viridis::viridis,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  reference_docx = NULL
)

create_geography_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  detailed_IUCN_redlist = NULL,
  report_kind = "static",
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  export_data = FALSE,
  output_file = NULL,
  output_dir = NULL,
  table_font_size = 10,
  ggtheme = NULL,
  separate_figure_folder = TRUE,
  value_on_fig = FALSE,
  endemic_species_per_region = NULL,
  accepted_species_per_region = NULL,
  tree_species_per_region = NULL,
  reference_docx = NULL
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{collection}{The name of the living collection}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{PlantClassification}{plant classification information}

\item{min_year}{The minimum year for analyses over time.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{report_kind}{The find of report to create, either \code{static} or \code{interactive}.}

\item{output_file}{The file path of the native report.}

\item{output_dir}{The output directory}

\item{export_data}{Flag (TRUE/FALSE) for whether to create a \code{.rda} and \code{.xlsx} files containing the data used to create the figures in the plot.}

\item{save_excel}{Flag (TRUE/FALSE) if export_data == TRUE, save excel allows for a 'xlsx' version of the data to be created or not. Note that for particualarly large collections, saving an excel version may cause the function to fail.}

\item{separate_figure_folder}{Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.}

\item{table_font_size}{Font size for tables.}

\item{ggtheme}{gg plot theme to be applied to all ggplots, see \code{ggthemes} package for pre-set themes.}

\item{value_on_fig}{Flag (TRUE/FALSE) for whether to include values of static plots.}

\item{interactive_colour}{colours for interactive report, default is viridis::viridis.}

\item{scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned}{scales for ggplot. Default is to use viridis.}

\item{reference_docx}{path to a .docx file whose style (design) the report copies.}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} package.}

\item{detailed_IUCN_redlist}{A data frame containing the history and geography of IUCN redlist information.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{endemic_species_per_region}{A data frame detailing the number of accepted endemic species found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}

\item{accepted_species_per_region}{A data frame detailing the number of accepted species (unique records in WCVP, including subspecies, etc) found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}

\item{tree_species_per_region}{A data frame detailing the number of accepted trees (unique records in WCVP, including subspecies, etc) found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}
}
\description{
\strong{Taxonomic Diversity}:Taxonomic presentation and composition in a collection. \strong{Geography}: Geographic representation of a collection.
}
