% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-diversity.R
\name{create_taxonomic_diversity_report}
\alias{create_taxonomic_diversity_report}
\alias{create_geography_report}
\title{Create diversity module}
\usage{
create_taxonomic_diversity_report(
  enriched_report,
  collection = NULL,
  wcvp = NULL,
  PlantClassification = BGSmartR::Diversity_classification$Classification,
  min_year = 1970,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  output_dir = NULL
)

create_geography_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  wcvp = NULL,
  detailed_IUCN_redlist = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  output_dir = NULL,
  endemic_species_per_region = NULL,
  accepted_species_per_region = NULL,
  tree_species_per_region = NULL,
  do_download = TRUE
)
}
\arguments{
\item{enriched_report}{The enriched report of a living collection (created in BGsmartR)}

\item{collection}{The name of the living collection}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}

\item{PlantClassification}{plant classification information}

\item{min_year}{The minimum year for analyses over time.}

\item{earliest_allowable_record}{Number, corresponding to the minimum allowable year in the "AccYear" column.}

\item{old_accession_year_codes}{Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on \code{earliest_allowable_record}.}

\item{output_dir}{The output directory}

\item{wgsrpd3}{World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from \code{rWCVP} package.}

\item{detailed_IUCN_redlist}{A data frame containing the history and geography of IUCN redlist information.}

\item{native}{string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.}

\item{extinct}{Flag (TRUE/FALSE) for whether to include extinct geographic locations.}

\item{doubtful_locations}{Flag (TRUE/FALSE) for whether to include doubtful locations.}

\item{endemic_species_per_region}{A data frame detailing the number of accepted endemic species found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}

\item{accepted_species_per_region}{A data frame detailing the number of accepted species (unique records in WCVP, including subspecies, etc) found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}

\item{tree_species_per_region}{A data frame detailing the number of accepted trees (unique records in WCVP, including subspecies, etc) found in WCVP for each region in wgsrpd3. (if running multiple reports saves having to do the calculation each time)}

\item{do_download}{A flag (TRUE/FALSE) for whether the download buttons are included.}
}
\description{
\strong{Taxonomic Diversity}:Taxonomic presentation and composition in a collection. \strong{Geography}: Geographic representation of a collection.
}
