% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{known_not_in_wcvp}
\alias{known_not_in_wcvp}
\title{known_not_in_wcvp()}
\usage{
known_not_in_wcvp(taxon_names)
}
\arguments{
\item{taxon_names}{A vector of taxon names}
}
\value{
A vector of the indices in taxon_names that can be removed.
}
\description{
A function that finds the taxon names which are known not to be in POWO. In particular this includes:
\itemize{
\item includes \code{ sp.}
\item includes \code{gx}
\item includes \code{'XX'} for some text XX. This is common notation for XX.
\item includes \code{[}
\item Begins with \code{Indet}. This is an indetminant and therefore won't be in POWO.
}
}
