% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrichment.R
\name{match_original_to_wcvp}
\alias{match_original_to_wcvp}
\title{Match report to POWO via taxon name}
\usage{
match_original_to_wcvp(original_report, wcvp)
}
\arguments{
\item{original_report}{A gardens original report}

\item{wcvp}{POWO database}
}
\value{
A list of length two containing:
\verb{$match} the index of the match from \code{taxon_names} to \code{wcvp}, where the match goes to the record with accepted status.
\verb{$message} a message detailing the match.
}
\description{
Match report to POWO via taxon name
}
