% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{shorten_message}
\alias{shorten_message}
\title{shorten_message}
\usage{
shorten_message(messages)
}
\arguments{
\item{messages}{messages about matching}
}
\value{
shorted form of messages
}
\description{
shorten_message
}
