% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{shorten_message}
\alias{shorten_message}
\title{Shorten message detailing match}
\usage{
shorten_message(messages)
}
\arguments{
\item{messages}{messages about matching}
}
\value{
shorted form of messages
}
\description{
Shorten message detailing match
}
